FileManagerとVegetableRepositoryFileManagerにわけてFirstScreenとそれ以外で分離
