+++ app/src/main/java/com/moritoui/vegegrowthapp/ui/FirstScreenViewModel.kt
@@ -1,6 +1,9 @@
 package com.moritoui.vegegrowthapp.ui
 
+import android.content.Context
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.moritoui.vegegrowthapp.model.FileManager
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -11,9 +14,25 @@ data class FirstScreenUiState(
     val inputText: String = ""
 )
 
-class FirstScreenViewModel : ViewModel() {
+class FirstScreenViewModel(
+    applicationContext: Context
+) : ViewModel() {
+    private val fileManger: FileManager
     private val _uiState = MutableStateFlow(FirstScreenUiState())
     val uiState: StateFlow<FirstScreenUiState> = _uiState.asStateFlow()
+
+    class FirstScreenFactory(
+        private val applicationContext: Context
+    ) : ViewModelProvider.Factory {
+        @Suppress("unchecked_cast")
+        override fun <T : ViewModel> create(modelClass: Class<T>) =
+            FirstScreenViewModel(
+                applicationContext
+            ) as T
+    }
+    init {
+        fileManger = FileManager(applicationContext = applicationContext)
+    }
 
     fun updateState(
         isOpenDialog: Boolean = _uiState.value.isOpenDialog,
